requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12.2
Flask-SSLify==0.1.5
gitdb2==2.0.2
GitPython==2.1.5
itsdangerous==0.24
Jinja2==2.9.6
Markdown==2.6.8
markdown2==2.3.4
MarkupSafe==1.0
mysqlclient==1.3.10
packaging==16.8
pyparsing==2.2.0
python-dotenv==0.6.4
PyYAML==3.12
six==1.10.0
smmap2==2.0.3
SQLAlchemy==1.1.11
Werkzeug==0.12.2
